[Graph Generator] Parallelize the generation algorithm (#23)
* initialize DeviceMemoryManagerWrapper

* add device memory offset for device tensors

* convert dfs to bfs

* parallelize search

* minor fix

* fix fingerprint issue

* reconstruct graphs for new states to avoid concurrency issue

* rename

* fix map key type

---------

+++ include/mirage/kernel/device_tensor.h
@@ -129,7 +129,7 @@ public:
   // offset in device memory
   int64_t fp_offset;
 
-  static int next_guid;
+  static int64_t next_guid;
 };
 
 NLOHMANN_DEFINE_TYPE_NON_INTRUSIVE(
